requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.1.7
django-extensions==2.2.1
django-fernet-fields==0.6
factory-boy==2.12.0
Faker==2.0.0
pycparser==2.19
PyMySQL==0.9.2
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
text-unidecode==1.2
